bug fix, moving patch into images
