+++ sparsecoding/transforms/__init__.py
@@ -1,7 +1,7 @@
-from .patch import quilt, patchify, sample_random_patches
 from .whiten import whiten, compute_whitening_stats
-from .images import whiten_images, compute_image_whitening_stats, WhiteningTransform
+from .images import whiten_images, compute_image_whitening_stats, WhiteningTransform, \
+                    quilt, patchify, sample_random_patches
 
 __all__ = ['quilt', 'patchify', 'sample_random_patches', 'whiten',
            'compute_whitening_stats', 'compute_image_whitening_stats',
-            'WhiteningTransform', 'whiten_images']
+           'WhiteningTransform', 'whiten_images']
